requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
djangorestframework==3.8.2
idna==2.6
pyTelegramBotAPI==3.6.3
pytz==2018.4
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
